Split interactions package
