+++ src/main/kotlin/dev/minn/jda/ktx/interactions/components/textinput.kt
@@ -15,7 +15,7 @@
  */
 
 
-package dev.minn.jda.ktx.interactions
+package dev.minn.jda.ktx.interactions.components
 
 import net.dv8tion.jda.api.interactions.components.text.TextInput
 import net.dv8tion.jda.api.interactions.components.text.TextInputStyle
