+++ src/main/kotlin/dev/minn/jda/ktx/interactions/components/utils.kt
@@ -14,28 +14,13 @@
  * limitations under the License.
  */
 
-package dev.minn.jda.ktx.interactions
+package dev.minn.jda.ktx.interactions.components
 
 import net.dv8tion.jda.api.entities.*
 import net.dv8tion.jda.api.interactions.commands.CommandInteractionPayload
 import net.dv8tion.jda.api.interactions.commands.OptionMapping
 import net.dv8tion.jda.api.interactions.commands.OptionType
-import java.io.File
 
-inline fun <reified T> optionType() = when(T::class.java) {
-    java.lang.Float::class.java, java.lang.Double::class.java -> OptionType.NUMBER
-    Integer::class.java, java.lang.Long::class.java, java.lang.Short::class.java, java.lang.Byte::class.java -> OptionType.INTEGER
-    String::class.java -> OptionType.STRING
-    User::class.java, Member::class.java -> OptionType.USER
-    Role::class.java -> OptionType.ROLE
-    java.lang.Boolean::class.java -> OptionType.BOOLEAN
-    File::class.java, Message.Attachment::class.java -> OptionType.ATTACHMENT
-    else -> when {
-        Channel::class.java.isAssignableFrom(T::class.java) -> OptionType.CHANNEL
-        IMentionable::class.java.isAssignableFrom(T::class.java) -> OptionType.MENTIONABLE
-        else -> OptionType.UNKNOWN
-    }
-}
 
 /**
  * Get an option value by name and resolve it to the desired type.
